Fix bugs in display methods for some C++ type classes. Change JSONArray type class in C++ to allow non-generic specification (e.g. for old-style tabled data) Update examples ascii_io & formatted_io3 to reflect that arrays are no longer assumed to be reallocatable.

+++ yggdrasil/languages/CPP/YggInterface.hpp
@@ -226,7 +226,7 @@ public:
   int send(const int nargs, ...) {
     va_list_t va;
     va_start(va.va, nargs);
-    int ret = vyggSend(_pi, nargs, va);
+    int ret = vyggSend(_pi, (size_t)nargs, va);
     va_end(va.va);
     return ret;
   }
